day10 JPA 검색 페이징 처리

+++ src/main/java/web/model/repository/ImgRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import web.model.entity.ImgEntity;
 
 @Repository
-public interface ImgRepository extends JpaRepository< ImgEntity, Integer > {
+public interface ImgRepository extends JpaRepository< ImgEntity, Long > {
 }
